set store state instead of component state
