+++ src/components/store.js
@@ -17,7 +17,15 @@ function mungeSources(sources) {
 
 function store(state, action) {
   if (!state) {
-    state = { sources: sources, pages: pages, text: '', file: '', url:'' };
+    state = {
+      sources: sources,
+      pages: pages,
+      text: '',
+      file: '',
+      url: '',
+      header: '',
+      header_description: '',
+      header_links: [] };
   }
   switch (action.type) {
   case "@@router/INIT_PATH":
@@ -34,6 +42,8 @@ function store(state, action) {
     return Object.assign({}, state, {text: action.text || '', file: action.file});
   case 'URL':
     return Object.assign({}, state, {url: action.url || ''});
+  case 'UPDATE_IDS':
+    return Object.assign({}, state, {header: action.header || '', header_description: action.header_description || '', header_links: action.header_links});
   default:
     return state;
   }
